Tests: Make daemon fixtures available to plugin packages
The `conftest.py` provides useful fixtures for tests that require a
started or stopped daemon client:

 * `started_daemon_client`: make sure daemon is running and return client
 * `stopped_daemon_client`: make sure daemon is stopped and return client
 * `submit_and_await`: submit a process or builder and wait for it to
   reach a particular process state.

These fixtures are not just useful for tests of `aiida-core` but also for
plugin packages. Therefore they are moved to the module
`aiida.manage.tests.pytest_fixtures` such that they can be easily imported
into the `conftest.py` of a plugin package.
