+++ aiida/manage/tests/pytest_fixtures.py
@@ -8,25 +8,24 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 # pylint: disable=redefined-outer-name,unused-argument
-"""
-Collection of pytest fixtures using the TestManager for easy testing of AiiDA plugins.
+"""Collection of pytest fixtures using the TestManager for easy testing of AiiDA plugins."""
+from __future__ import annotations
 
- * aiida_profile
- * aiida_profile_clean
- * aiida_profile_clean_class
- * aiida_localhost
- * aiida_local_code_factory
-
-"""
 import asyncio
+import pathlib
 import shutil
 import tempfile
+import time
 
+import plumpy
 import pytest
 
 from aiida.common.log import AIIDA_LOGGER
 from aiida.common.warnings import warn_deprecation
+from aiida.engine import Process, ProcessBuilder, submit
+from aiida.engine.daemon.client import DaemonClient
 from aiida.manage.tests import get_test_backend_name, get_test_profile_name, test_manager
+from aiida.orm import ProcessNode
 
 
 @pytest.fixture(scope='function')
@@ -240,3 +239,81 @@ def aiida_local_code_factory(aiida_localhost):
         return code.store()
 
     return get_code
+
+
+@pytest.fixture(scope='session')
+def daemon_client(aiida_profile):
+    """Return a daemon client for the configured test profile for the test session.
+
+    The daemon will be automatically stopped at the end of the session.
+    """
+    daemon_client = DaemonClient(aiida_profile._manager._profile)  # pylint: disable=protected-access
+
+    try:
+        yield daemon_client
+    finally:
+        daemon_client.stop_daemon(wait=True)
+        assert not daemon_client.is_daemon_running
+
+
+@pytest.fixture()
+def started_daemon_client(daemon_client):
+    """Ensure that the daemon is running for the test profile and return the associated client."""
+    if not daemon_client.is_daemon_running:
+        daemon_client.start_daemon()
+        assert daemon_client.is_daemon_running
+
+    yield daemon_client
+
+
+@pytest.fixture()
+def stopped_daemon_client(daemon_client):
+    """Ensure that the daemon is not running for the test profile and return the associated client."""
+    if daemon_client.is_daemon_running:
+        daemon_client.stop_daemon(wait=True)
+        assert not daemon_client.is_daemon_running
+
+    yield daemon_client
+
+
+@pytest.fixture
+def submit_and_await(started_daemon_client):
+    """Submit a process and wait for it to achieve the given state."""
+
+    def _factory(
+        submittable: Process | ProcessBuilder | ProcessNode,
+        state: plumpy.ProcessState = plumpy.ProcessState.FINISHED,
+        timeout: int = 5
+    ):
+        """Submit a process and wait for it to achieve the given state.
+
+        :param submittable: A process, a process builder or a process node. If it is a process or builder, it is
+            submitted first before awaiting the desired state.
+        :param state: The process state to wait for, by default it waits for the submittable to be ``FINISHED``.
+        :param timeout: The time to wait for the process to achieve the state.
+        :raises RuntimeError: If the process fails to achieve the specified state before the timeout expires.
+        """
+        if not isinstance(submittable, ProcessNode):
+            node = submit(submittable)
+        else:
+            node = submittable
+
+        start_time = time.time()
+
+        while node.process_state is not state:
+
+            if node.is_excepted:
+                raise RuntimeError(f'The process excepted: {node.exception}')
+
+            if time.time() - start_time >= timeout:
+                daemon_log_file = pathlib.Path(started_daemon_client.daemon_log_file).read_text(encoding='utf-8')
+                daemon_status = 'running' if started_daemon_client.is_daemon_running else 'stopped'
+                raise RuntimeError(
+                    f'Timed out waiting for process with state `{node.process_state}` to enter state `{state}`.\n'
+                    f'Daemon <{started_daemon_client.profile.name}|{daemon_status}> log file content: \n'
+                    f'{daemon_log_file}'
+                )
+
+        return node
+
+    return _factory
